refactor(ui): make GridView and AppImageView more generic and reusable
- Refactored GridView to support dynamic data and customizable layouts
- Updated AppImageView to handle different image sources and styles
- Improved code readability and maintainability for both components
- Ensured backward compatibility with existing UI elements
