+++ SnapShop-Admin/Views/CustomViews/AppGrids/ProductsGrid.swift
@@ -1,0 +1,26 @@
+//
+//  ProductsGrid.swift
+//  SnapShop-Admin
+//
+//  Created by Hadir on 27/05/2024.
+//
+
+import SwiftUI
+
+struct ProductsGrid: View {
+    
+    var productsList: [Product]
+    
+    var body: some View {
+        GridView(items: productsList, columns: 2) { product in
+            ProductsCustomCell(productImageURL: product.image?.src ?? "",
+                               productName: product.title ?? "",
+                               productCategory: product.productType ?? "",
+                               productPrice: "\(product.variants?.first?.price ?? "") EGP")
+        }
+    }
+}
+
+
+
+
